requirements.txt
django==5.0.6
django-environ==0.11.2
django-allauth
djangorestframework==3.15.2
rest-framework-simplejwt==5.0.0
django-cors-headers==4.0.0
